Revert ABI fixed width types (#773)
Reverts villagesql/villagesql-server#755

It seems we should investigate this more, perhaps when we tackle
Windows.

We need to understand whether this is actually a problem or if this is
unnecessary.

Created by bonsai from 29157a5e440438ab72a49079acd554b3fd5cba9f

+++ villagesql/sdk/include/villagesql/abi/preview/statement_event.h
@@ -106,7 +106,7 @@ typedef struct {
   // Query text. Not null-terminated; use query_len.
   // Lifetime: copy before return.
   const char *query;
-  uint64_t query_len;
+  size_t query_len;
 
   // Authenticated user name, or empty string.
   // Lifetime: connection lifetime.
